-> begin process of splitting out SA verification from accounts.views
